Gestion des poules d'un championnat
[NEW] Gestion de plusieurs poules dans un championnat (scraper +
frontend). Ajout d'un menu déroulant pour sélectionner la poule qui est
enregistrée en local.
[NEW] Sur la page de l'équipe, possibilité de filtrer les matchs
(domicile, extérieur ou tous)
