+++ scraper/config.py
@@ -1,2 +1,2 @@
-championship_url = 'https://resultats.ffbb.com/championnat/b5e6211fb466.html'
-pool_id = 'b5e621216574'
+championship_url = 'https://resultats.ffbb.com/championnat/b5e6211fe70c.html'
+pool_id = 'b5e62121bf85'
