+++ team/index.html
@@ -1,77 +1,103 @@
 <!DOCTYPE html>
 <html lang="fr">
+<head>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1" />
+	<title>FFBB Alternative</title>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.9.1/chart.min.js"></script>
+	<link rel="mask-icon" href="../assets/ffbb_logo.svg" color="#5bbad5">
+	<link rel="stylesheet" href="../css/style.css"/>
+</head>
 
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>FFBB Alternative</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.9.1/chart.min.js"></script>
-    <script src="script.js"></script>
-    <link rel="mask-icon" href="../assets/ffbb_logo.svg" color="#5bbad5">
-    <link rel="stylesheet" href="../style.css"/>
-  </head>
+<body>
+	<div class="header">
+		<div class="header-infos">
+			<a href="../" class="header-back"><svg viewBox="0 0 24 24"><path d="M13 8.5L9.5 12l3.5 3.5"></path><path d="M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10z"></path></svg></a>
+			<div class="header-texts">
+				<div class="header-subtitle"></div>
+				<div class="header-title"></div>
+			</div>
+		</div>
+	</div>
 
-  <body>
-    <div class="header">
-      <div class="header-infos">
-        <a href="../" class="header-back"><svg viewBox="0 0 24 24"><path d="M13 8.5L9.5 12l3.5 3.5"></path><path d="M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10z"></path></svg></a>
-        <div class="header-texts">
-          <div class="header-subtitle"></div>
-          <div class="header-title"></div>
-        </div>
-      </div>
-    </div>
+	<div class="container">
+		<div class="charts-header">
+			<div class="title">Matchs</div>
+			<div class="charts-mode">
+				<label class="charts-radio">
+					<input type="radio" name="games-selection" id="tous" checked="checked">
+					<span>Tous</span>
+					<span class="charts-checkmark">
+						<svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
+					</span>
+				</label>
+				<label class="charts-radio">
+					<input type="radio" name="games-selection" id="domicile">
+					<span>Domicile</span>
+					<span class="charts-checkmark">
+						<svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
+					</span>
+				</label>
+				<label class="charts-radio">
+					<input type="radio" name="games-selection" id="exterieur">
+					<span>Extérieur</span>
+					<span class="charts-checkmark">
+						<svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
+					</span>
+				</label>
+			</div>
+		</div>
+	</div>
 
-    <div class="container">
-      <div class="title">Statistiques</div>
-      <div class="stats"></div>
-    </div>
+	<div class="container">
+		<div class="title">Statistiques</div>
+		<div class="stats"></div>
+	</div>
 
-    <div class="container">
-      <div class="charts-header">
-        <div class="title">Graphiques</div>
-        <div class="charts-mode">
-          <label class="charts-radio">
-            <input type="radio" name="charts-mode-selection" id="combine" checked="checked">
-            <span>Combinées</span>
-            <span class="charts-checkmark">
-              <svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
-            </span>
-          </label>
-          <label class="charts-radio">
-            <input type="radio" name="charts-mode-selection" id="divide">
-            <span>Divisées</span>
-            <span class="charts-checkmark">
-              <svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
-            </span>
-          </label>
-        </div>
-      </div>
-      <div class="charts"></div>
-    </div>
+	<div class="container">
+		<div class="charts-header">
+			<div class="title">Graphiques</div>
+			<div class="charts-mode">
+				<label class="charts-radio">
+					<input type="radio" name="charts-mode-selection" id="combine" checked="checked">
+					<span>Combinées</span>
+					<span class="charts-checkmark">
+						<svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
+					</span>
+				</label>
+				<label class="charts-radio">
+					<input type="radio" name="charts-mode-selection" id="divide">
+					<span>Divisées</span>
+					<span class="charts-checkmark">
+						<svg><path d="M9.6 1c.2.3.3.7 0 1l-5 5.3c-.6.7-1.5.7-2.2.1l-2-1.9a.6.6 0 0 1 .9-.9l2 1.9c.1.1.3.1.4 0l5-5.4c.2-.3.6-.3.9 0Z"></path></svg>
+					</span>
+				</label>
+			</div>
+		</div>
+		<div class="charts"></div>
+	</div>
 
-    <div class="container">
-      <div class="title">Résultats</div>
-      <div class="fixtures" id="results_fixtures"></div>
-    </div>
-    
-    <div class="container">
-      <div class="title">Calendrier</div>
-      <div class="fixtures" id="calendar_fixtures"></div>
-    </div>
+	<div class="container">
+		<div class="title">Résultats</div>
+		<div class="fixtures" id="results_fixtures"></div>
+	</div>
+	
+	<div class="container">
+		<div class="title">Calendrier</div>
+		<div class="fixtures" id="calendar_fixtures"></div>
+	</div>
 
-    <div class="footer">
-      <div class="copyright">
-        <a href="https://lucasboucher.fr"><img src="../assets/lb_logo.svg" alt="Lucas Boucher logo"></a>
-        <div>© 2022 Lucas Boucher</div>
-      </div>
-      <ul class="links">
-        <!-- <li><a href="../dev">/dev</a></li> -->
-        <li class="hide-mobile"><a href="https://github.com/lucasboucher/FFBB_Alternative">GitHub</a></li>
-        <li class="hide-mobile"><a href="http://www.ffbb.com/accueil">FFBB</a></li>
-        <li class="hide-mobile"><a href="https://onefootball.com/">Onefootball</a></li>
-      </ul>
-    </div>
-  </body>
+	<div class="footer">
+		<div class="copyright hidden-on-mobile">
+			<div>© 2022 <a href="https://lucasboucher.fr">Lucas Boucher</a></div>
+		</div>
+		<ul class="links">
+			<li><a href="https://github.com/lucasboucher/FFBB_Alternative">GitHub</a></li>
+			<li><a href="http://www.ffbb.com/accueil">FFBB</a></li>
+		</ul>
+	</div>
 
+	<script src="../js/team.js"></script>
+
+</body>
 </html>
